requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==5.1.7
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
isort==6.0.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
sqlparse==0.5.3
stripe==12.0.0
drf-yasg==1.21.10
celery==5.5.2
redis==5.2.1
django-celery-beat==2.8.0
